Button changes not working but organized components into folders
