+++ src/App.css
@@ -7,16 +7,10 @@
   pointer-events: none;
 }
 
-@media (prefers-reduced-motion: no-preference) {
-  .App-logo {
-    animation: App-logo-spin infinite 20s linear;
-  }
-}
-
 .App-header {
   /* The following was changed from default to make the site look modern -AX */
-  background-color: #f2f2f2;
-  color: #333;
+  background-color: #db1010;
+  color: #1118a0;
   min-height: 100vh;
   padding: 20px;
   display: flex;
