+++ src/components/ButtonComponent/ButtonComponent.module.css
@@ -1,0 +1,92 @@
+/* module.css files ensure that component isolation and multiple styles are a possibility. */
+
+.fancyButton,
+.fancyButton *,
+.fancyButton :after,
+.fancyButton :before,
+.fancyButton:after,
+.fancyButton:before {
+  border: 0 solid;
+  box-sizing: border-box;
+}
+.fancyButton {
+  -webkit-tap-highlight-color: transparent;
+  -webkit-appearance: button;
+  background-color: #000;
+  background-image: none;
+  color: #fff;
+  font-family: ui-sans-serif, system-ui, -apple-system, BlinkMacSystemFont,
+    Segoe UI, Roboto, Helvetica Neue, Arial, Noto Sans, sans-serif,
+    Apple Color Emoji, Segoe UI Emoji, Segoe UI Symbol, Noto Color Emoji;
+  font-size: 100%;
+  font-weight: 900;
+  line-height: 1.5;
+  margin: 0;
+  -webkit-mask-image: none; /* Adjusted as radial-gradient doesn't seem necessary */
+  padding: 0;
+  text-transform: uppercase;
+}
+.fancyButton:disabled {
+  cursor: default;
+}
+.fancyButton:-moz-focusring {
+  outline: auto;
+}
+.fancyButton svg {
+  vertical-align: middle;
+}
+.fancyButton [hidden] {
+  display: none;
+}
+.fancyButton {
+  --thickness: 0.3rem;
+  --roundness: 1.2rem;
+  --color: #eff6ff;
+  --opacity: 0.6;
+  -webkit-backdrop-filter: blur(100px);
+  backdrop-filter: blur(100px);
+  background: none;
+  background: hsla(0, 0%, 100%, 0.2);
+  border: none;
+  border-radius: var(--roundness);
+  color: var(--color);
+  cursor: pointer;
+  display: block;
+  font-family: Poppins, "sans-serif";
+  font-size: 1rem;
+  font-weight: 500;
+  padding: 0.8rem 3rem;
+  position: relative;
+}
+.fancyButton:hover {
+  background: hsla(0, 0%, 100%, 0.3);
+  filter: brightness(1.2);
+}
+.fancyButton:active {
+  --opacity: 0;
+  background: hsla(0, 0%, 100%, 0.1);
+}
+.fancyButton svg {
+  border-radius: var(--roundness);
+  display: block;
+  filter: none; /* Adjusted as there's no filter with id `#glow` defined */
+  height: 100%;
+  left: 0;
+  position: absolute;
+  top: 0;
+  width: 100%;
+}
+.fancyButton rect {
+  fill: none;
+  stroke: var(--color);
+  stroke-width: var(--thickness);
+  rx: var(--roundness);
+  stroke-linejoin: round;
+  stroke-dasharray: 185%;
+  stroke-dashoffset: 80;
+  -webkit-animation: none; /* Adjusted as there's no animation called `snake` */
+  animation: none; /* Adjusted as there's no animation called `snake` */
+  height: 100%;
+  opacity: var(--opacity);
+  width: 100%;
+}
